serialization

+++ src/metrics/cpu_ram.rs
@@ -2,9 +2,9 @@ use sysinfo::{CpuRefreshKind, MemoryRefreshKind, RefreshKind, System};
 
 pub struct CpuRamMetrics {
     sys: System,
-    pub cpu_usage: f64,
+    pub cpu_usage: f32,
     pub cpu_name: String,
-    pub cpu_frequency: u64,
+    pub cpu_frequency: u32,
     pub total_ram: u64,
     pub used_ram: u64,
     pub is_supported: bool,
@@ -23,10 +23,10 @@ impl CpuRamMetrics{
                 .with_ram())
             );
         let cpu_name = sys.cpus()[0].brand().to_string();
-        let cpu_frequency = sys.cpus()[0].frequency();
+        let cpu_frequency = sys.cpus()[0].frequency() as u32;
 
-        let total_ram = sys.total_memory();
-        let used_ram = sys.used_memory();
+        let total_ram = sys.total_memory() as u64;
+        let used_ram = sys.used_memory() as u64;
 
         let is_supported = sysinfo::IS_SUPPORTED_SYSTEM;
         Self{
@@ -46,8 +46,8 @@ impl CpuRamMetrics{
         self.update_cpu_usage();
 
         self.sys.refresh_memory();
-        self.total_ram = self.sys.total_memory();
-        self.used_ram = self.sys.used_memory();
+        self.total_ram = self.sys.total_memory() as u64;
+        self.used_ram = self.sys.used_memory() as u64;
 
     }
     fn update_cpu_usage(&mut self){
@@ -57,10 +57,10 @@ impl CpuRamMetrics{
         let mut sum = 0.0;
         let mut count = 0;
         for cpu in self.sys.cpus() {
-            sum += cpu.cpu_usage() as f64;
+            sum += cpu.cpu_usage() as f32;
             count += 1;
         }
-        self.cpu_usage = sum / count as f64;
+        self.cpu_usage = sum / count as f32;
 
     }
 }
